--- a/pyproject.toml
+++ b/pyproject.toml
@@ -53,6 +53,9 @@ scripts.tinytext = "tinytext.cli:main"
 [tool.hatch]
 version.source = "vcs"
 
+[tool.hatch.build.hooks.vcs]
+version-file = "src/tinytext/_version.py"
+
 [tool.hatch.version.raw-options]
 local_scheme = "no-local-version"