With Python 3.12.4:

```sh
python -X importtime -c "import tinytext" 2> import.log && tuna import.log
```

Cuts out the 22ms `importlib.metadata` import.

# `main`

<img width="1624" alt="image" src="https://github.com/user-attachments/assets/042a2bf7-a842-4ec2-93f2-6ede3a8f1451">

```console
❯ hyperfine -w 8 "tinytext --version"
Benchmark 1: tinytext --version
  Time (mean ± σ):      49.2 ms ±   3.7 ms    [User: 38.2 ms, System: 9.5 ms]
  Range (min … max):    45.1 ms …  66.4 ms    62 runs
```

# PR

<img width="1624" alt="image" src="https://github.com/user-attachments/assets/ef4c7669-cb92-4038-bb06-3b18537b7926">

```console
❯ hyperfine -w 8 "tinytext --version"
Benchmark 1: tinytext --version
  Time (mean ± σ):      26.4 ms ±   1.0 ms    [User: 19.8 ms, System: 5.5 ms]
  Range (min … max):    24.6 ms …  29.1 ms    105 runs
```